client-hints/accept_ch-manual.https.html

<html>
<title>Accept-CH-Lifetime manual test</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id=test>
    <p>This is a manual test, since the persistent client hints tests require
    the browser to have a cache implementation for storing client hints
    preferences of different origins. Such a cache implementation may only be
    available only when the full implementation of browser is run.

    The test additionally opens another html web page. One test is run in this
    web page, and another in the second web page.
</div>

<script>

// This test fetches resources/accept_ch.html. The response headers to
// that webpage contains only the Accept-CH header. Due to the missing
// Accept-CH-Lifetime header, the user-agent should not persist origin
// preferences for the client hints specified in Accept-CH header.

// Next, to verify that the origin preferences were not persisted by the user
// agent, this test fetches resources/do_not_expect_client_hints_headers.html
// in a new window. Fetching of
// resources/do_not_expect_client_hints_headers.html
// verifies that the user agent does not send the client hints in the request
// headers.

// Verify the initial state to make sure that the browser does not have client
// hints preferences cached from a previous run of the test.
promise_test(t => {
  return fetch("echo_client_hints_received.py").then(r => {
    assert_equals(r.status, 200)
    // Verify that the browser did not include client hints in the request
    // headers when fetching echo_client_hints_received.py.
    assert_false(r.headers.has("device-memory-received"), "device-memory-received");
  });
}, "Test that the browser does not have client hints preferences cached");

promise_test(t => {
  return fetch("resources/accept_ch.html").then(r => {
    assert_equals(r.status, 200)
    assert_false(r.headers.has("device-memory-received"), "device-memory-received");

    // Open a new window.
    window.open("resources/do_not_expect_client_hints_headers.html");
  });
}, "Test receiving Accept-CH header");


</script>

</body>
</html>
